Force healthcheck to run using ipv4
Due to this issue: https://github.com/denoland/deno/issues/5269
Deno is not guaranteed to bind to ipv6 as well as ipv4 when
binding to '0.0.0.0'. Equally it is not guaranteed to bind to ipv4
when binding to '::'. So we have
to bind ipv4 and force the check to be ipv4 inside the
container.

This changes the check to force ipv4.

This appears to manifest only on certain devices (in this case a recent Apple Silicon Mac).